+++ app/Helpers/EmailVerificationHelper.php
@@ -3,7 +3,7 @@
 namespace App\Helpers;
 
 use App\Auth\EmailVerification;
-use App\User;
+use App\Models\User;
 
 class EmailVerificationHelper {
     public static function checkVerificationAttempt($userID, $token)
