+++ app/Http/Controllers/AuthController.php
@@ -9,7 +9,7 @@ use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\Controller;
 use App\LoginSession;
 use App\Rules\CurrentPasswordMatches;
-use App\User;
+use App\Models\User;
 use DateTime;
 use Hash;
 use Illuminate\Support\Facades\Request as RequestFacade;
