+++ app/Utils/InfluxDB/InfluxDBWrapperInterface.php
@@ -2,7 +2,7 @@
 
 namespace App\Utils\InfluxDB;
 
-use App\Speedtest;
+use App\Models\Speedtest;
 
 interface InfluxDBWrapperInterface
 {
