+++ tests/Unit/Helpers/SpeedtestHelper/PercentageThresholdTest.php
@@ -4,7 +4,7 @@ namespace Tests\Unit\Helpers\SpeedtestHelper;
 
 use App\Helpers\SettingsHelper;
 use App\Helpers\SpeedtestHelper;
-use App\Speedtest;
+use App\Models\Speedtest;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
